Add buffer size emission configuration to JackOutput module
Added a configuration option to the JackOutput module that allows for the emission of the used buffer size whenever there is a change. This new 'emit_buffer_size' parameter was added to facilitate the monitoring of buffer usage, thereby improving efficiency. Furthermore, the JackOutput class now inherits from the BasicEventProducer class in order to facilitate this event emission.